requirements.txt
assemblyai
google-ai-generativelanguage
google-generativeai
gtts
huggingface-hub
langchain
langchain-community
langchain-core
langchain-google-genai
numpy
opencv-python
pandas
pillow
plotly
pycountry
pydantic
python-dotenv
requests
requests-oauthlib
requests-toolbelt
SpeechRecognition
streamlit
supabase


